// src/app/types/job.ts
export type Job = {
  id: string;
  jobTitle: string;
  companyName: string;
  location: string;
  industryType: string;
  department: string;
  employmentType: string;
  roleCategory: string;
  jobDescription: string;
  jobHighlights: string[];
  keySkills: string[];
  educationRequirements: string;
  experienceLevel: string;
  salaryRange: string;
  shortDescription: string;
};